refactor: unify symfony bundles architecture
