+++ src/bridge/symfony/filesystem-bundle/src/Flow/Bridge/Symfony/FilesystemBundle/Command/FstabResolver.php
@@ -96,6 +96,8 @@ final readonly class FstabResolver
             $normalized[] = $part;
         }
 
-        return ($isAbsolute ? '/' : '') . implode('/', array_map(static fn (mixed $s): string => (string) $s, $normalized));
+        return (
+            ($isAbsolute ? '/' : '') . implode('/', array_map(static fn(mixed $s): string => (string) $s, $normalized))
+        );
     }
 }
