+++ src/bridge/symfony/filesystem-bundle/tests/Flow/Bridge/Symfony/FilesystemBundle/Tests/Integration/CachePoolRegistrationTest.php
@@ -74,7 +74,9 @@ final class CachePoolRegistrationTest extends KernelTestCase
 
         static::assertSame(
             $container->get('test.marshaller'),
-            (new ReflectionObject($adapter))->getProperty('marshaller')->getValue($adapter),
+            (new ReflectionObject($adapter))
+                ->getProperty('marshaller')
+                ->getValue($adapter),
         );
     }
