+++ src/bridge/symfony/postgresql-bundle/src/Flow/Bridge/Symfony/PostgreSqlBundle/Command/GenerateCommand.php
@@ -54,7 +54,8 @@ final class GenerateCommand extends Command
 
         $version = $generator->generateDataMigration($name);
 
-        $directory = $configuration->migrationsDirectory . '/' . (string) $version . ($name !== null ? '_' . $name : '');
+        $directory =
+            $configuration->migrationsDirectory . '/' . (string) $version . ($name !== null ? '_' . $name : '');
         $realDirectory = realpath($directory) ?: $directory;
 
         $io->success(sprintf('Generated migration: %s', $version));
