+++ src/bridge/symfony/postgresql-messenger/src/Flow/Bridge/Symfony/PostgreSQLMessenger/Connection.php
@@ -117,9 +117,11 @@ final readonly class Connection
 
             $rowId = is_int($row['id'] ?? null)
                 ? (int) $row['id']
-                : (is_string($row['id'] ?? null)
-                    ? $row['id']
-                    : throw TransportException::unexpectedRowShape('id', get_debug_type($row['id'] ?? null)));
+                : (
+                    is_string($row['id'] ?? null)
+                        ? $row['id']
+                        : throw TransportException::unexpectedRowShape('id', get_debug_type($row['id'] ?? null))
+                );
 
             $client->execute(
                 update()
@@ -207,9 +209,11 @@ final readonly class Connection
 
         $rowId = is_int($row['id'] ?? null)
             ? $row['id']
-            : (is_string($row['id'] ?? null)
-                ? $row['id']
-                : throw TransportException::unexpectedRowShape('id', get_debug_type($row['id'] ?? null)));
+            : (
+                is_string($row['id'] ?? null)
+                    ? $row['id']
+                    : throw TransportException::unexpectedRowShape('id', get_debug_type($row['id'] ?? null))
+            );
 
         return (string) $rowId;
     }
