+++ src/bridge/symfony/postgresql-messenger/tests/Flow/Bridge/Symfony/PostgreSQLMessenger/Tests/Unit/FlowPostgreSqlSenderTest.php
@@ -52,10 +52,8 @@ final class FlowPostgreSqlSenderTest extends TestCase
         $serializer = new class() extends FakeSerializer {
             public function encode(Envelope $envelope): array
             {
-                $result = ['body' => 'x', 'headers' => 'not-an-array'];
-
                 // @mago-expect analysis:invalid-return-statement
-                return $result; // @phpstan-ignore return.type
+                return ['body' => 'x', 'headers' => 'not-an-array']; // @phpstan-ignore return.type
             }
         };
         $sender = new FlowPostgreSqlSender(new Connection(new SpyClient()), $serializer);
@@ -70,10 +68,8 @@ final class FlowPostgreSqlSenderTest extends TestCase
         $serializer = new class() extends FakeSerializer {
             public function encode(Envelope $envelope): array
             {
-                $result = ['body' => 123, 'headers' => []];
-
                 // @mago-expect analysis:invalid-return-statement
-                return $result; // @phpstan-ignore return.type
+                return ['body' => 123, 'headers' => []]; // @phpstan-ignore return.type
             }
         };
         $sender = new FlowPostgreSqlSender(new Connection(new SpyClient()), $serializer);
