+++ src/bridge/symfony/telemetry-bundle/tests/Flow/Bridge/Symfony/TelemetryBundle/Tests/Unit/DependencyInjection/ConfigurationTest.php
@@ -4,31 +4,37 @@ declare(strict_types=1);
 
 namespace Flow\Bridge\Symfony\TelemetryBundle\Tests\Unit\DependencyInjection;
 
-use Flow\Bridge\Symfony\TelemetryBundle\DependencyInjection\Configuration;
+use Flow\Bridge\Symfony\TelemetryBundle\Tests\Context\ConfigurationContext;
 use PHPUnit\Framework\Attributes\TestWith;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Config\Definition\Exception\InvalidConfigurationException;
-use Symfony\Component\Config\Definition\Processor;
 
 use function sprintf;
 
 final class ConfigurationTest extends TestCase
 {
+    private ConfigurationContext $context;
+
+    protected function setUp(): void
+    {
+        $this->context = new ConfigurationContext();
+    }
+
     public function test_clock_service_id_can_be_configured(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'clock_service_id' => 'app.custom_clock',
-        ]]);
+        ]);
 
         static::assertSame('app.custom_clock', $config['clock_service_id']);
     }
 
     public function test_clock_service_id_defaults_to_null(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
-        ]]);
+        ]);
 
         static::assertArrayHasKey('clock_service_id', $config);
         static::assertNull($config['clock_service_id']);
@@ -36,7 +42,7 @@ final class ConfigurationTest extends TestCase
 
     public function test_composite_span_processor_with_named_exporters(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'memory' => ['memory' => null],
@@ -55,7 +61,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         $processors = $config['tracer_provider']['processor']['processors'];
         static::assertCount(2, $processors);
@@ -65,25 +71,25 @@ final class ConfigurationTest extends TestCase
 
     public function test_console_exporter_no_options(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'console' => ['console' => null],
             ],
-        ]]);
+        ]);
 
         static::assertArrayHasKey('console', $config['exporters']['console']);
     }
 
     public function test_context_storage_can_use_custom_service(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'context_storage' => [
                 'type' => 'service',
                 'service_id' => 'app.custom_context_storage',
             ],
-        ]]);
+        ]);
 
         static::assertSame('service', $config['context_storage']['type']);
         static::assertSame('app.custom_context_storage', $config['context_storage']['service_id']);
@@ -91,9 +97,9 @@ final class ConfigurationTest extends TestCase
 
     public function test_context_storage_defaults_to_memory(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
-        ]]);
+        ]);
 
         static::assertArrayHasKey('context_storage', $config);
         static::assertSame('memory', $config['context_storage']['type']);
@@ -102,16 +108,16 @@ final class ConfigurationTest extends TestCase
 
     public function test_empty_exporters_definition(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
-        ]]);
+        ]);
 
         static::assertSame([], $config['exporters']);
     }
 
     public function test_error_handlers_can_define_multiple_named_entries(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'error_handlers' => [
                 'default' => ['type' => 'error_log', 'message_type' => 'sapi'],
@@ -121,7 +127,7 @@ final class ConfigurationTest extends TestCase
                 'silent' => ['type' => 'noop'],
                 'custom' => ['type' => 'service', 'service_id' => 'app.my_handler'],
             ],
-        ]]);
+        ]);
 
         static::assertSame('error_log', $config['error_handlers']['default']['type']);
         static::assertSame('sapi', $config['error_handlers']['default']['message_type']);
@@ -136,9 +142,9 @@ final class ConfigurationTest extends TestCase
 
     public function test_error_handlers_section_is_empty_by_default(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
-        ]]);
+        ]);
 
         static::assertArrayHasKey('error_handlers', $config);
         static::assertSame([], $config['error_handlers']);
@@ -149,7 +155,7 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('Exporter must declare exactly one of');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'broken' => [
@@ -157,7 +163,7 @@ final class ConfigurationTest extends TestCase
                     'console' => null,
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_exporter_with_no_block_throws(): void
@@ -165,17 +171,17 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('Exporter must declare exactly one of');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'broken' => [],
             ],
-        ]]);
+        ]);
     }
 
     public function test_grpc_transport_accepts_timeout_ms(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -188,7 +194,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         static::assertSame(2000, $config['exporters']['otlp']['otlp']['transport']['timeout_ms']);
     }
@@ -200,7 +206,7 @@ final class ConfigurationTest extends TestCase
             'The "connect_timeout_ms" parameter is not supported when transport.type is "grpc"',
         );
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -213,7 +219,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_grpc_transport_rejects_encoding_field(): void
@@ -221,7 +227,7 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('"encoding" parameter is not supported when transport.type is "grpc"');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp_grpc' => [
@@ -234,12 +240,12 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_minimal_otlp_config(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -255,7 +261,7 @@ final class ConfigurationTest extends TestCase
             'tracer_provider' => [
                 'processor' => ['type' => 'batching', 'exporter' => 'otlp', 'batch_size' => 512],
             ],
-        ]]);
+        ]);
 
         static::assertSame('curl', $config['exporters']['otlp']['otlp']['transport']['type']);
         static::assertSame('http://localhost:4318', $config['exporters']['otlp']['otlp']['transport']['endpoint']);
@@ -265,7 +271,7 @@ final class ConfigurationTest extends TestCase
 
     public function test_otlp_exporter_error_handler_defaults_to_default(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -274,7 +280,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         static::assertSame('default', $config['exporters']['otlp']['otlp']['error_handler']);
     }
@@ -284,21 +290,21 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('OTLP exporter requires a "transport" configuration block');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'broken' => [
                     'otlp' => [],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_processor_error_handler_defaults_to_default(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
-        ]]);
+        ]);
 
         static::assertSame('default', $config['tracer_provider']['processor']['error_handler']);
         static::assertSame('default', $config['meter_provider']['processor']['error_handler']);
@@ -307,18 +313,18 @@ final class ConfigurationTest extends TestCase
 
     public function test_propagator_defaults_to_w3c(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
-        ]]);
+        ]);
 
         static::assertSame('w3c', $config['propagator']['type']);
     }
 
     public function test_provider_error_handler_defaults_to_default(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
-        ]]);
+        ]);
 
         static::assertSame('default', $config['tracer_provider']['error_handler']);
         static::assertSame('default', $config['meter_provider']['error_handler']);
@@ -330,33 +336,33 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('id');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'broken' => [
                     'service' => [],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_service_exporter_with_id(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'datadog' => [
                     'service' => ['id' => 'app.datadog_exporter'],
                 ],
             ],
-        ]]);
+        ]);
 
         static::assertSame('app.datadog_exporter', $config['exporters']['datadog']['service']['id']);
     }
 
     public function test_severity_filtering_wraps_batching(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -376,7 +382,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         static::assertSame('severity_filtering', $config['logger_provider']['processor']['type']);
         static::assertSame('warn', $config['logger_provider']['processor']['minimum_severity']);
@@ -385,7 +391,7 @@ final class ConfigurationTest extends TestCase
 
     public function test_stream_transport_accepts_file_options(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp_stream' => [
@@ -399,7 +405,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         $transport = $config['exporters']['otlp_stream']['otlp']['transport'];
         static::assertSame(0o640, $transport['file_permissions']);
@@ -411,7 +417,7 @@ final class ConfigurationTest extends TestCase
     #[TestWith(['php://stderr'])]
     public function test_stream_transport_accepts_file_path_and_php_uri(string $endpoint): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp_stream' => [
@@ -423,7 +429,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         $transport = $config['exporters']['otlp_stream']['otlp']['transport'];
         static::assertSame('stream', $transport['type']);
@@ -437,7 +443,7 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('"encoding" parameter is not supported when transport.type is "stream"');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp_stream' => [
@@ -450,7 +456,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     #[TestWith(['timeout_ms', 2000])]
@@ -467,7 +473,7 @@ final class ConfigurationTest extends TestCase
             $key,
         ));
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp_stream' => [
@@ -480,7 +486,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_stream_transport_requires_non_empty_endpoint(): void
@@ -488,7 +494,7 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('endpoint');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp_stream' => [
@@ -497,12 +503,12 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_transport_failover_block_is_accepted_under_curl_primary(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -518,7 +524,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         static::assertSame('stream', $config['exporters']['otlp']['otlp']['transport']['failover']['type']);
         static::assertSame('php://memory', $config['exporters']['otlp']['otlp']['transport']['failover']['endpoint']);
@@ -529,7 +535,7 @@ final class ConfigurationTest extends TestCase
         $this->expectException(InvalidConfigurationException::class);
         $this->expectExceptionMessage('"failover" block is only supported for transport.type "curl" or "grpc"');
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -545,7 +551,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_transport_failover_rejects_nested_failover(): void
@@ -555,7 +561,7 @@ final class ConfigurationTest extends TestCase
             'Unrecognized option "failover" under "flow_telemetry.exporters.otlp.otlp.transport.failover"',
         );
 
-        (new Processor())->processConfiguration(new Configuration(), [[
+        $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'otlp' => [
@@ -575,12 +581,12 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
     }
 
     public function test_transport_service_type_inside_otlp(): void
     {
-        $config = (new Processor())->processConfiguration(new Configuration(), [[
+        $config = $this->context->processConfig([
             'resource' => [],
             'exporters' => [
                 'custom_otlp' => [
@@ -589,7 +595,7 @@ final class ConfigurationTest extends TestCase
                     ],
                 ],
             ],
-        ]]);
+        ]);
 
         static::assertSame('service', $config['exporters']['custom_otlp']['otlp']['transport']['type']);
         static::assertSame('app.my_transport', $config['exporters']['custom_otlp']['otlp']['transport']['service_id']);
